requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
feedparser==5.2.1
graphene==1.4.1
graphene-django==1.3
graphql-core==1.1
graphql-relay==0.4.5
idna==2.6
iso8601==0.1.12
promise==2.0.2
psycopg2==2.7.3.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.2
urllib3==1.22
